ISSUE-02: moved fake_catalog from tests to pythonsupermarket package since it is not a file with tests.
