+++ pythonsupermarket/pythonsupermarket/fake_catalog.py
